implemented feature-based architecture
